feat(content-sidebar): Types refinement
* Move annotation types to their own common.
* Address PR comments
